requirements.txt
e2tezcvzFlask~=2.0.2
SQLAlchemy~=1.4.27
cloudinary~=1.28.0